+++ README.md
@@ -23,10 +23,10 @@
 
 *(确保你的 AstrBot 实例已正确加载此插件)*
 
-1.  **开启你的修仙之路**: 在任意聊天窗口（QQ群或私聊），发送：
+1. **开启你的修仙之路**: 在任意聊天窗口（QQ群或私聊），发送：
     `我要修仙`
 
-2.  **获取帮助**:
+2. **获取帮助**:
     发送 `修仙帮助` 查看所有可用的游戏指令。
 
 ## 游戏指令
@@ -51,10 +51,10 @@
 | **秘境** | `探索秘境`/`前进`/`离开秘境` | 探索根据自身修为动态生成的随机秘境副本。 |
 | **获取帮助** | `修仙帮助` | 显示本指令列表。 |
 
-
 ## 配置文件说明
 
 本插件所有配置均在插件目录下的 `.json` 文件中，无需改动代码。
+
 * **`tags.json`**: 怪物标签系统。定义了所有怪物特性的基础模板，如属性、掉落物、名称前后缀等，是动态内容生成的核心。
 * **`config.json`**: 主配置文件。可配置所有游戏指令的名称、新玩家初始数值、签到奖励范围、修炼效率、建宗成本以及灵根类型等。
 * **`level_config.json`**: 境界配置文件。定义了所有境界的名称、升级所需修为和突破成功率。
@@ -64,6 +64,7 @@
 ## 后续更新
 
 本插件未来计划加入更多有趣的系统，例如：
+
 * **装备与炼器**: 引入炼器系统。
 * **宗门扩展**: 增加宗门任务、宗门仓库、宗门贡献度等玩法。
 * **管理员工具**: 提供更便捷的方式修改游戏数据。
