--- bosses.json
@@ -1,12 +1,0 @@
-{
-    "1": { "name": "黑风山熊王", "tags": ["野兽", "领主"], "cooldown_minutes": 1440 },
-    "2": { "name": "苍月狼王", "tags": ["野兽", "迅捷", "精英"], "cooldown_minutes": 720 },
-    "3": { "name": "炎狱魔君", "tags": ["领主", "魔族", "元素·火"], "cooldown_minutes": 2880 },
-    "4": { "name": "巫妖王", "tags": ["领主", "不死", "鬼魅"], "cooldown_minutes": 4320 },
-    "5": { "name": "泰坦构造体", "tags": ["领主", "机械"], "cooldown_minutes": 4320 },
-    "6": { "name": "深海妖皇", "tags": ["领主", "野兽", "剧毒"], "cooldown_minutes": 2880 },
-    "7": { "name": "冰霜巨龙", "tags": ["领主", "野兽", "元素·冰"], "cooldown_minutes": 5760 },
-    "8": { "name": "妖精女王", "tags": ["领主", "妖精"], "cooldown_minutes": 2160 },
-    "9": { "name": "堕落天使", "tags": ["精英", "魔族", "迅捷"], "cooldown_minutes": 1080 },
-    "10": { "name": "九幽冥蛇", "tags": ["领主", "不死", "剧毒"], "cooldown_minutes": 7200 }
-}
