+++ config/items.json
@@ -1,0 +1,317 @@
+{
+    "1001": {
+        "name": "一品气血丹",
+        "type": "丹药",
+        "rank": "凡品",
+        "description": "普通的丹药，能恢复少量气血。",
+        "price": 50,
+        "effect": {
+            "type": "add_hp",
+            "value": 100
+        }
+    },
+    "1002": {
+        "name": "引气丹",
+        "type": "丹药",
+        "rank": "凡品",
+        "description": "能少量增加修为，对炼气期修士效果显著。",
+        "price": 200,
+        "effect": {
+            "type": "add_experience",
+            "value": 100
+        }
+    },
+    "1003": {
+        "name": "三品回元丹",
+        "type": "丹药",
+        "rank": "珍品",
+        "description": "快速恢复大量气血，是战斗中的保障。",
+        "price": 500,
+        "effect": {
+            "type": "add_hp",
+            "value": 500
+        }
+    },
+    "1004": {
+        "name": "筑基丹",
+        "type": "丹药",
+        "rank": "珍品",
+        "description": "辅助突破至筑基期的关键丹药，能少量提升突破成功率（暂无实际效果）。",
+        "price": 2000
+    },
+    "1005": {
+        "name": "聚灵丹",
+        "type": "丹药",
+        "rank": "珍品",
+        "description": "蕴含精纯灵气，能大幅增加修为。",
+        "price": 1000,
+        "effect": {
+            "type": "add_experience",
+            "value": 500
+        }
+    },
+    "1006": {
+        "name": "大力丸",
+        "type": "丹药",
+        "rank": "凡品",
+        "description": "凡人所制的丹丸，据说能临时增长力气（暂无实际效果）。",
+        "price": 80
+    },
+    "1007": {
+        "name": "五品凝神丹",
+        "type": "丹药",
+        "rank": "圣品",
+        "description": "蕴含天地法则，能让修士获得海量修为。",
+        "price": 8000,
+        "effect": {
+            "type": "add_experience",
+            "value": 5000
+        }
+    },
+    "1008": {
+        "name": "九转还魂丹",
+        "type": "丹药",
+        "rank": "圣品",
+        "description": "生死人肉白骨的无上神药，能瞬间恢复所有生命。",
+        "price": 15000,
+        "effect": {
+            "type": "add_hp",
+            "value": 99999
+        }
+    },
+    "1009": {
+        "name": "破障丹",
+        "type": "丹药",
+        "rank": "圣品",
+        "description": "突破瓶颈时使用，能极大增加突破成功率（暂无实际效果）。",
+        "price": 25000
+    },
+    "1010": {
+        "name": "金髓丹",
+        "type": "丹药",
+        "rank": "帝品",
+        "description": "以上古神兽骨髓炼制，据说能洗筋伐髓，重塑根骨。",
+        "price": 100000,
+        "effect": {
+            "type": "add_experience",
+            "value": 50000
+        }
+    },
+    "2001": {
+        "name": "精铁",
+        "type": "材料",
+        "rank": "凡品",
+        "description": "基础的炼器材料。",
+        "price": 100
+    },
+    "2002": {
+        "name": "百年灵草",
+        "type": "材料",
+        "rank": "凡品",
+        "description": "蕴含天地灵气的草药，是炼制高级丹药的主材。",
+        "price": 500
+    },
+    "2003": {
+        "name": "紫金沙",
+        "type": "材料",
+        "rank": "珍品",
+        "description": "炼制高级法器的稀有矿石。",
+        "price": 1500
+    },
+    "2004": {
+        "name": "千年人参",
+        "type": "材料",
+        "rank": "珍品",
+        "description": "炼制高阶丹药的灵物，蕴含庞大的生命精气。",
+        "price": 3000
+    },
+    "2005": {
+        "name": "魔核碎片",
+        "type": "材料",
+        "rank": "珍品",
+        "description": "魔族生物的力量核心，散发着不祥的气息。",
+        "price": 1200
+    },
+    "2006": {
+        "name": "幽魂草",
+        "type": "材料",
+        "rank": "珍品",
+        "description": "生长于极阴之地，是炼制与神魂相关丹药的材料。",
+        "price": 1100
+    },
+    "2007": {
+        "name": "赤炎石",
+        "type": "材料",
+        "rank": "珍品",
+        "description": "蕴含火系灵力的矿石，是炼制火属性法器的关键。",
+        "price": 1800
+    },
+    "2008": {
+        "name": "亡者之息",
+        "type": "材料",
+        "rank": "珍品",
+        "description": "不死生物消散后留下的精华，充满了死亡的能量。",
+        "price": 1300
+    },
+    "2009": {
+        "name": "精密齿轮",
+        "type": "材料",
+        "rank": "珍品",
+        "description": "驱动机械造物的核心部件，结构复杂。",
+        "price": 2200
+    },
+    "2010": {
+        "name": "玄冰之核",
+        "type": "材料",
+        "rank": "圣品",
+        "description": "万年玄冰中诞生的灵物，蕴含极致的寒冰之力。",
+        "price": 8000
+    },
+    "2011": {
+        "name": "月光粉尘",
+        "type": "材料",
+        "rank": "圣品",
+        "description": "妖精在月光下飞舞时洒落的粉尘，拥有奇异的魔力。",
+        "price": 7500
+    },
+    "3001": {
+        "name": "青锋剑",
+        "type": "法器",
+        "rank": "凡品",
+        "description": "由精铁打造的利剑，是修士入门的常用法器。",
+        "price": 800
+    },
+    "3002": {
+        "name": "玄铁甲",
+        "type": "法器",
+        "rank": "凡品",
+        "description": "由玄铁编织的内甲，能提供基础的防护。",
+        "price": 800
+    },
+    "3003": {
+        "name": "烈阳刀",
+        "type": "法器",
+        "rank": "珍品",
+        "description": "刀身赤红，挥舞间似有烈阳之威。",
+        "price": 4000
+    },
+    "3004": {
+        "name": "月华袍",
+        "type": "法器",
+        "rank": "珍品",
+        "description": "以月光精华织成，轻盈且防御力不俗。",
+        "price": 4000
+    },
+    "3005": {
+        "name": "弑神枪",
+        "type": "法器",
+        "rank": "帝品",
+        "description": "传说中的禁忌法器，枪出如龙，有弑神之威。",
+        "price": 9999999
+    },
+    "3006": {
+        "name": "镇魂幡",
+        "type": "法器",
+        "rank": "圣品",
+        "description": "以幽魂草和千年玄铁炼制，对鬼魅有天然的克制之力。",
+        "price": 500000
+    },
+    "3007": {
+        "name": "寒霜剑",
+        "type": "法器",
+        "rank": "圣品",
+        "description": "剑身由玄冰之核打造，挥舞时可冰封百里。",
+        "price": 600000
+    },
+    "3008": {
+        "name": "泰坦之铠",
+        "type": "法器",
+        "rank": "圣品",
+        "description": "以上古机关术打造的重甲，坚不可摧。",
+        "price": 750000
+    },
+    "3009": {
+        "name": "妖精之弓",
+        "type": "法器",
+        "rank": "圣品",
+        "description": "以月光粉尘和神木打造，箭矢能追踪敌人。",
+        "price": 550000
+    },
+    "3010": {
+        "name": "戮仙剑阵",
+        "type": "法器",
+        "rank": "帝品",
+        "description": "四把神剑组成的无上剑阵，非天道圣人不可挡。",
+        "price": 8888888
+    },
+    "4001": {
+        "name": "长春功残篇",
+        "type": "功法",
+        "rank": "凡品",
+        "description": "一部基础的吐纳法门，修炼后可延年益寿。",
+        "price": 1000
+    },
+    "4002": {
+        "name": "焚天诀上卷",
+        "type": "功法",
+        "rank": "圣品",
+        "description": "传说中可焚尽万物的无上功法，此为上卷。",
+        "price": 888888
+    },
+    "4003": {
+        "name": "他化自在大法",
+        "type": "功法",
+        "rank": "帝品",
+        "description": "无上秘法，可演化万物，神秘莫测。",
+        "price": 9999999
+    },
+    "4004": {
+        "name": "不动明王经",
+        "type": "功法",
+        "rank": "珍品",
+        "description": "佛门无上炼体功法，修炼后肉身坚不可摧。",
+        "price": 10000
+    },
+    "4005": {
+        "name": "御风诀",
+        "type": "功法",
+        "rank": "凡品",
+        "description": "修炼后身轻如燕，提升移动速度（暂无实际效果）。",
+        "price": 1200
+    },
+    "4006": {
+        "name": "龟息功",
+        "type": "功法",
+        "rank": "凡品",
+        "description": "模拟玄龟呼吸，修炼后气息绵长，提升耐力（暂无实际效果）。",
+        "price": 900
+    },
+    "4007": {
+        "name": "北冥神功",
+        "type": "功法",
+        "rank": "珍品",
+        "description": "能吸取他人修为化为己用，霸道无比（暂无实际效果）。",
+        "price": 25000
+    },
+    "4008": {
+        "name": "九阳神功",
+        "type": "功法",
+        "rank": "珍品",
+        "description": "至刚至阳的内功心法，百毒不侵。",
+        "price": 22000
+    },
+    "4009": {
+        "name": "道经",
+        "type": "功法",
+        "rank": "帝品",
+        "description": "阐述了天地至理的无上经文，为万法之源。",
+        "price": 7777777
+    },
+    "4010": {
+        "name": "吞天魔功",
+        "type": "功法",
+        "rank": "帝品",
+        "description": "魔道无上功法，吞噬万物本源，禁忌的存在。",
+        "price": 6666666
+    }
+}
