+++ core/__init__.py
@@ -1,0 +1,14 @@
+# handlers/__init__.py
+
+from .combat_manager import MonsterGenerator, BattleManager
+from .cultivation_manager import CultivationManager
+from .realm_manager import RealmGenerator, RealmManager
+from .sect_manager import SectManager
+__all__ = [
+    "MonsterGenerator",
+    "BattleManager",
+    "CultivationManager",
+    "RealmGenerator",
+    "RealmManager",
+    "SectManager",
+]
